moved implemented Position object where is supposed to be
